+++ CadAddinManager/ViewModel/AddInManagerViewModel.cs
@@ -242,7 +242,7 @@ public class AddInManagerViewModel : ViewModelBase
         return MainTrees;
     }
 
-    private void ExecuteAddinCommandClick()
+    public void ExecuteAddinCommandClick()
     {
         try
         {